refactor(frontend): reorganize remaining utils
