+++ src/AnthropicTool/TeXLinterFixAgent.ts
@@ -6,11 +6,11 @@ import { ToolResult } from './base';
 
 // Local imports - types
 import { ValidationResult } from './types';
-import { LinterMessage } from '../utils/linterUtils';
+import { LinterMessage } from '../frontend/latex/linter';
 
 // Local imports - utils
 import * as workspaceFileUtils from '../utils/files';
-import * as linterUtils from '../utils/linterUtils';
+import * as linterUtils from '../frontend/latex/linter';
 
 // Local imports - Logging
 import * as logger from '../logger/logUtils';
