+++ src/agent/utils/userVars.ts
@@ -4,7 +4,7 @@ import * as path from 'path';
 
 import { AgentLogger } from '../../logger/AgentLogger';
 import { readFile } from '../../utils/files';
-import { setVarFromFile } from '../../utils/fileVarUtils';
+import { setVarFromFile } from '../../frontend/files/vars';
 import { getXmlFormatFromFiles, getListOfFiles } from '../../utils/promptUtils';
 import { AgentConfig } from '../core/AgentConfig';
 import { AgentSetting } from '../core/AgentDataclass';
