+++ src/commands/fileSelectionCommands.ts
@@ -5,14 +5,11 @@ import * as vscode from 'vscode';
 import * as logger from '../logger/logUtils';
 
 // Local imports - utilities
-import {
-  showInfoMessage,
-  showErrorMessage,
-} from '../frontend-utils/commonUtils';
+import { showInfoMessage, showErrorMessage } from '../frontend/ui/messageUtils';
 import { getRelativePath } from '../utils/files';
-import { listInputFiles } from '../frontend-utils/fileListingUtils';
+import { listInputFiles } from '../frontend/files/listing';
 import { getIncludedExtensions } from '../utils/fileTypeUtils';
-import { selectFile, selectFiles } from '../utils/files';
+import { selectFile, selectFiles } from '../frontend/files/dialog';
 const CHANNEL = 'fileSelectionCommands';
 logger.initialize(CHANNEL);
