+++ src/extension.ts
@@ -3,10 +3,10 @@ import * as vscode from 'vscode';
 
 // Local imports - core
 import * as logger from './logger/logUtils';
-import { initializeSecrets } from './utils/secretUtils';
-import { copyDefaultAgents, configureLatexSettings } from './utils/setupUtils';
+import { initializeSecrets } from './frontend/secrets';
+import { copyDefaultAgents, configureLatexSettings } from './frontend/setup';
 import { watchConfig } from './utils/config';
-import { disposeDiffRefresh } from './utils/diffViewUtils';
+import { disposeDiffRefresh } from './frontend/ui/diffView';
 
 // Local imports - components
 import { ProgressViewProvider } from './progressView/ProgressViewProvider';
