+++ src/frontend/ui/diffView.ts
@@ -2,8 +2,8 @@
 import * as vscode from 'vscode';
 
 // Local imports - log
-import * as logger from '../logger/logUtils';
-import { REFRESH_THRESHOLD_MS } from './config';
+import * as logger from '../../logger/logUtils';
+import { REFRESH_THRESHOLD_MS } from '../../utils/config';
 
 const CHANNEL = 'DiffRefresh';
 logger.initialize(CHANNEL);
