+++ src/latex/textConnection.ts
@@ -7,10 +7,7 @@ import { formatProviderError } from '../utils/sdkErrorUtils';
 
 // Local imports - log
 import * as logger from '../logger/logUtils';
-import {
-  getApiKey as getSecretApiKey,
-  ApiProvider,
-} from '../utils/secretUtils';
+import { getApiKey as getSecretApiKey, ApiProvider } from '../frontend/secrets';
 
 const CHANNEL = 'LaTeXCommands';
 logger.initialize(CHANNEL);
