+++ src/progressView/ProgressViewContentProvider.ts
@@ -5,7 +5,7 @@ import * as vscode from 'vscode';
 
 // Local imports - log
 import * as logger from '../logger/logUtils';
-import { buildWebviewHtml } from '../utils/webviewHtmlUtils';
+import { buildWebviewHtml } from '../frontend/webview/html';
 
 const CHANNEL = 'Webview';
 logger.initialize(CHANNEL);
