+++ src/utils/files/index.ts
@@ -1,4 +1,2 @@
 export * from './workspaceFileUtils';
 export * from './absoluteFileUtils';
-export * from './pathUtils';
-export * from './fileDialogUtils';
